changed e to event, and moved zoomScaleSliderLeft and zoomScaleSliderRight to constants
